Crate massa_api_exports

Structs

AddressInfo

All you ever dream to know about an address

BlockInfo

refactor to delete

BlockInfoContent

Block content

BlockSummary

A block resume (without the block itself)

CompactAddressInfo

Less information about an address

DatastoreEntryInput

Datastore entry query input structure

DatastoreEntryOutput

Datastore entry query output structure

EndorsementInfo

All you wanna know about an endorsement

ExecuteReadOnlyResponse

The response to a request for a read-only execution.

LedgerInfo

Current balance ledger info

NodeStatus

node status

OperationInfo

Operation and contextual info about it

OperationInput

operation input

ReadOnlyBytecodeExecution

read only bytecode execution request

ReadOnlyCall

read SC call request

RollsInfo

Roll counts

SlotAmount

slot / amount pair

TimeInterval

Just a wrapper with a optional beginning and end

Enums

ListType

Bootsrap lists types

ReadOnlyResult

The result of the read-only execution.

ScrudOperation

SCRUD operations